Fixed bugs, added help menu

+++ main.py
@@ -1,5 +1,5 @@
-from random import randint
 from plot import Plot
+from utils.help import ShowHelp
 from utils.colors import COLORS
 from utils.window import Window
 import pickle
@@ -106,7 +106,7 @@ $$$$$$$$/  $$$$$$/   $$$$$$$ |$$/ $$/   $$/
         Dashboard(username)
 
 def ViewAll(username: str):
-    Header("📖  VIEW ALL ENTRIES  📖")
+    Header("📖  VIEW ALL ENTRIES  📖", color=COLORS.LIGHT_GREEN)
     screen.print("All of Diary entries: ", color=COLORS.YELLOW)
     screen.print()
     idx = 1
@@ -161,12 +161,14 @@ def ViewAll(username: str):
         screen.render()
 
 def ViewDateRange(username: str):
-    Header("📖  VIEW ENTRIES BY DATE 📖")
-    screen.print("Choose how you’d like to view your entries:")
+    Header("📖  VIEW ENTRIES BY DATE 📖", color=COLORS.LIGHT_GREEN)
+    screen.print("Choose how you’d like to view your entries:", color=COLORS.YELLOW)
     screen.print()
-    start = datetime.strptime(screen.input("Enter Start Date: (DD-MM-YYYY) "), '%d-%m-%Y')
-    end = datetime.strptime(screen.input("Enter End Date: (DD-MM-YYYY) "), '%d-%m-%Y')
+    screen.line()
+    start = datetime.strptime(screen.input("Enter Start Date: (DD-MM-YYYY) ", color=COLORS.LIGHT_GRAY), '%d-%m-%Y')
+    end = datetime.strptime(screen.input("Enter End Date: (DD-MM-YYYY) ", color=COLORS.LIGHT_GRAY), '%d-%m-%Y')
     screen.print()
+    screen.line()
     idx = 1
     for i in os.listdir(f".lifelog/{username}"):
         with open(f".lifelog/{username}/{i}", 'rb') as file:
@@ -175,31 +177,34 @@ def ViewDateRange(username: str):
             date = datetime.strptime(datetime.strptime(data['date'], "%d %B %Y").strftime("%d-%m-%Y"), "%d-%m-%Y")
             content = data['content'][:30] + "..."
             if date > start and date < end:
-                screen.print(f"[{idx}] [{date.strftime('%d %b %Y')}] - [{mood} ]")
-                screen.print(f"Title: {i.replace('_', ' ').title()[:-4]}")
+                screen.print(f"[{idx}] Title: {i.replace('_', ' ').title()[:-4]}", color=COLORS.LIGHT_GREEN)
+                screen.print(f"[{date.strftime('%d %b %Y')}] - [{mood} ]", color=COLORS.YELLOW)
                 screen.line()
-                screen.print(content)
+                screen.print(content, color=COLORS.LIGHT_GRAY)
                 screen.line()
         idx += 1
         screen.print()
         screen.print()
         screen.render()
-    screen.input("Press ENTER to go back!!")
+    screen.line()
+    screen.input("Press ENTER to go back!!", color=COLORS.DARK_GRAY)
     screen.clear()
     screen.render()
 
 def ViewByMood(username: str):
-    Header("💭 VIEW ENTRIES BY MOOD 💭")
-    screen.print("Which mood would you like to explore?")
+    Header("💭 VIEW ENTRIES BY MOOD 💭", color=COLORS.LIGHT_GREEN)
+    screen.print("Which mood would you like to explore?", color=COLORS.YELLOW)
     screen.print()
-    screen.print("""[1] 😊 Happy
-[2] 😌 Calm
-[3] 😔 Sad
-[4] 😤 Stressed
-[5] 😎 Confident
-[6] Back to menu
-    """)
-    mood = int(screen.input("Enter your mood: "))
+    screen.line()
+    screen.print("> [1] 😊 Happy", color=COLORS.LIGHT_GREEN)
+    screen.print("[2] 😌 Calm", color=COLORS.LIGHT_GRAY)
+    screen.print("[3] 😔 Sad", color=COLORS.LIGHT_GRAY)
+    screen.print("[4] 😤 Stressed", color=COLORS.LIGHT_GRAY)
+    screen.print("[5] 😎 Confident", color=COLORS.LIGHT_GRAY)
+    screen.print("[6] Back to menu", color=COLORS.LIGHT_GRAY)
+    mood = screen.input("Enter your mood: ", color=COLORS.YELLOW)
+    if mood == '': mood = 1
+    mood = int(mood)
     if mood == 6:
         screen.clear()
         screen.render()
@@ -208,7 +213,7 @@ def ViewByMood(username: str):
         idx = 1
         screen.clear()
         screen.render()
-        Header(f"{mood[0]} ENTRIES WITH MOOD [{mood[1:]}] {mood[0]}")
+        Header(f"{mood[0]} ENTRIES WITH MOOD [{mood[2:]}] {mood[0]}", color=COLORS.LIGHT_GREEN)
         for i in os.listdir(f".lifelog/{username}"):
             with open(f".lifelog/{username}/{i}", 'rb') as file:
                 data = pickle.load(file)
@@ -216,66 +221,73 @@ def ViewByMood(username: str):
                 date = data['date']
                 content = data['content'][:30] + "..."
                 if mood == dMood:
-                    screen.print(f"[{idx}] [{date}] - Title: {i.replace('_', ' ').title()[:-4]}")
+                    screen.print(f"[{idx}] [{date}] - Title: {i.replace('_', ' ').title()[:-4]}", color=COLORS.YELLOW)
                     screen.line()
-                    screen.print(content)
+                    screen.print(content, color=COLORS.LIGHT_GRAY)
                     screen.line()
             idx += 1
             screen.print()
             screen.print()
             screen.render()
-        screen.input("Press ENTER to go back!!")
+        screen.input("Press ENTER to go back!!", color=COLORS.DARK_GRAY)
         screen.clear()
         screen.render()
-
-def ViewEntries(username: str):
-    Header("📖  VIEW PAST ENTRIES  📖")
-    screen.print("Choose how you’d like to view your entries:", color=COLORS.YELLOW)
-    screen.line()
-    screen.print()
-    screen.print("> [1] View all entries", color=COLORS.LIGHT_GREEN)
-    screen.print("[2] View by date range", color=COLORS.LIGHT_GRAY)
-    screen.print("[3] View by mood", color=COLORS.LIGHT_GRAY)
-    screen.print("[4] Back to dashboard", color=COLORS.LIGHT_GRAY)
-    screen.line()
-    choice = screen.input("Enter your choice (1-4): ", color=COLORS.YELLOW)
-    if choice == '':
-        choice = 1
-    choice = int(choice)
-    screen.clear()
-    screen.render()
-    if choice == 1:
-        ViewAll(username)
-
-    elif choice == 2:
-        ViewDateRange(username)
-
-    elif choice == 3:
         ViewByMood(username)
 
-    elif choice == 4:
+def ViewEntries(username: str):
+    Header("📖  VIEW PAST ENTRIES  📖", color=COLORS.LIGHT_GREEN)
+    if len(os.listdir(f".lifelog/{username}/")) == 0:
+        screen.print("You don't have any diary entries. Create one first", color=COLORS.YELLOW)
+        screen.input("Press ENTER to go back", color=COLORS.DARK_GRAY)
         screen.clear()
         screen.render()
     else:
+        screen.print("Choose how you’d like to view your entries:", color=COLORS.YELLOW)
+        screen.line()
+        screen.print()
+        screen.print("> [1] View all entries", color=COLORS.LIGHT_GREEN)
+        screen.print("[2] View by date range", color=COLORS.LIGHT_GRAY)
+        screen.print("[3] View by mood", color=COLORS.LIGHT_GRAY)
+        screen.print("[4] Back to dashboard", color=COLORS.LIGHT_GRAY)
+        screen.line()
+        choice = screen.input("Enter your choice (1-4): ", color=COLORS.YELLOW)
+        if choice == '':
+            choice = 1
+        choice = int(choice)
         screen.clear()
         screen.render()
+        if choice == 1:
+            ViewAll(username)
+
+        elif choice == 2:
+            ViewDateRange(username)
+
+        elif choice == 3:
+            ViewByMood(username)
+
+        elif choice == 4:
+            screen.clear()
+            screen.render()
+        else:
+            screen.clear()
+            screen.render()
 
 def MoodAnalytics(username: str):
-    Header("📊  MOOD ANALYTICS  📊")
-    screen.print("Your Mood record:")
-    screen.print()
+    Header("📊  MOOD ANALYTICS  📊", color=COLORS.LIGHT_GREEN)
+    screen.print("Your Mood record:", color=COLORS.YELLOW)
+    screen.line()
     stats = {1:0,2:0,3:0,4:0,5:0}
     for i in os.listdir(f".lifelog/{username}"):
         with open(f".lifelog/{username}/{i}", 'rb') as file:
             data = pickle.load(file)
             mood = int(data['mood'])
             stats[mood] += 1
-    screen.print(f"😊 Happy     →  {stats[1]} days")
-    screen.print(f"😌 Calm      →  {stats[2]} days")
-    screen.print(f"😔 Sad       →  {stats[3]} days")
-    screen.print(f"😤 Stressed  →  {stats[4]} days")
-    screen.print(f"😎 Confident →  {stats[5]} days")
-    screen.print()
+    screen.print(f"😊 Happy     →  {stats[1]} days", color=(COLORS.LIGHT_GREEN if stats[1] !=0 else COLORS.DARK_GRAY))
+    screen.print(f"😌 Calm      →  {stats[2]} days", color=(COLORS.YELLOW if stats[2] !=0 else COLORS.DARK_GRAY))
+    screen.print(f"😔 Sad       →  {stats[3]} days", color=(COLORS.LIGHT_PURPLE if stats[3] !=0 else COLORS.DARK_GRAY))
+    screen.print(f"😤 Stressed  →  {stats[4]} days", color=(COLORS.LIGHT_RED if stats[4] !=0 else COLORS.DARK_GRAY))
+    screen.print(f"😎 Confident →  {stats[5]} days", color=(COLORS.CYAN if stats[5] !=0 else COLORS.DARK_GRAY))
+    screen.line()
     if stats != {1:0,2:0,3:0,4:0,5:0}:
         max = 0
         mostCommon = 1
@@ -285,26 +297,27 @@ def MoodAnalytics(username: str):
                 i = max
         screen.print()
         screen.line()
-        screen.print(f"Most common mood: {moods[mostCommon]}")
+        screen.print(f"Most common mood: {moods[mostCommon]}", color=COLORS.LIGHT_GREEN)
         screen.line()
-        screen.print("Plotting chart....")
+        screen.print("Plotting chart....", color=COLORS.LIGHT_GRAY)
         screen.render()
         try:
             Plot(stats)
             pass
         except Exception as E:
-            screen.print("[x] Could not plot chart due to some unforseen reasons. Sorry")
+            screen.print("[x] Could not plot chart due to some unforseen reasons. Sorry", color=COLORS.LIGHT_RED)
             screen.print(str(E))
             screen.render()
             time.sleep(1)
         screen.render()
-    screen.input("Press ENTER to go back!!")
+    screen.input("Press ENTER to go back!!", color=COLORS.DARK_GRAY)
     screen.clear()
     screen.render()
 
 def EditDelete(username: str):
-    Header("🧹  EDIT / DELETE ENTRIES  🧹")
-    screen.print("Here are your recent entries:")
+    Header("🧹  EDIT / DELETE ENTRIES  🧹", color=COLORS.LIGHT_GREEN)
+    screen.print("Here are your recent entries:", color=COLORS.YELLOW)
+    screen.line()
     idx = 1
     titles = []
     for i in os.listdir(f".lifelog/{username}"):
@@ -313,56 +326,78 @@ def EditDelete(username: str):
             mood = moods[int(data['mood'])][2:]
             titles.append(i.replace("_", " ").title()[:-4])
             date = datetime.strptime(data['date'], "%d %B %Y").strftime("%d %b %Y")
-            screen.print(f"[{idx}] [{date}] - {i.replace('_', ' ').title()[:-4]} ({mood})")
+            screen.print(f"[{idx}] [{date}] - {i.replace('_', ' ').title()[:-4]} ({mood})", color=COLORS.LIGHT_GRAY)
         idx += 1
-    screen.line()
+    if titles == []:
+        screen.print("You don't have any Diary entries as of now!", color=COLORS.YELLOW)
+        screen.input("Press ENTER to go back", color=COLORS.DARK_GRAY)
+        screen.clear()
+        screen.render()
 
-    try:
-        choice = int(screen.input("Index of the entry to edit/delete (or ENTER to go back):"))-1
-        if choice < 0 or choice > idx:
-            screen.clear()
-            screen.render()
-        else:
-            screen.clear()
-            screen.render()
-            screen.print(f"You selected: {titles[choice]}")
-            screen.print()
-            screen.print("What would you like to do?")
-            screen.print("[1] Edit this entry")
-            screen.print("[2] Delete this entry")
-            screen.print("[3] Cancel")
-            screen.line()
-            action = int(screen.input("Enter your choice: "))
-            if action == 1:
-                screen.clear()
+    else:
+        try:
+            choice = int(screen.input("Index of the entry to edit/delete (or ENTER to go back):", color=COLORS.YELLOW))-1
+            if choice < 0 or choice > idx:
+                screen.print(f"[x] You only have {idx-1} diaries!", color=COLORS.LIGHT_RED)
                 screen.render()
-                screen.print("Editing Entry...")
-                newMood = screen.input("New Mood: ")
-                screen.print("Modify text below (type END when done):")
-                diary = screen.editor("> ") 
-                screen.print()
-                screen.line()
-                screen.print()
-                with open(f".lifelog/{username}/{titles[choice].replace(' ','_').lower()}.dat", 'wb') as file:
-                    data = { "content": diary, "mood": newMood, "date": datetime.now().strftime("%d %B %Y") }
-                    pickle.dump(data, file)
-                screen.print("[✔] Entry updated successfully!")
-                screen.input("Press ENTER to go back")
-                screen.clear()
-                screen.render()
-
-            elif action == 2:
-                confirm = screen.input("Are you sure you want to delete this entry? (Y/N)")
-                if confirm.upper() == 'Y':
-                    os.remove(f'.lifelog/{username}/{titles[choice].replace(" ", "_").lower()}.dat')
+                time.sleep(1)
                 screen.clear()
                 screen.render()
             else:
                 screen.clear()
                 screen.render()
-    except Exception:
-        screen.clear()
-        screen.render()
+                screen.print(f"TITLE: {titles[choice]}", color=COLORS.YELLOW)
+                screen.print()
+                screen.print("What would you like to do?", color=COLORS.YELLOW)
+                screen.line()
+                screen.print("> [1] Edit this entry", color=COLORS.LIGHT_GREEN)
+                screen.print("[2] Delete this entry", color=COLORS.LIGHT_GRAY)
+                screen.print("[3] Cancel", color=COLORS.LIGHT_GRAY)
+                screen.line()
+                action = screen.input("Enter your choice: ", color=COLORS.YELLOW)
+                if action == '':action =1
+                action = int(action)
+                if action == 1:
+                    screen.clear()
+                    screen.render()
+                    screen.print()
+                    screen.print(titles[choice], color=COLORS.LIGHT_GREEN, centered=True)
+                    screen.line()
+                    screen.print("Editing Entry...", color=COLORS.YELLOW)
+                    newMood = screen.input("New Mood: ", color=COLORS.YELLOW)
+                    if not newMood.isnumeric() or int(newMood) > 5 or int(newMood)<0:
+                        screen.print("[x] Invalid mood", color=COLORS.LIGHT_RED)
+                        screen.render()
+                        time.sleep(1)
+                        screen.clear()
+                        screen.render()
+                    else:
+                        screen.print("Modify text below (type END when done):", color=COLORS.LIGHT_GRAY)
+                        diary = screen.editor("> ", color=COLORS.DARK_GRAY)
+                        screen.print()
+                        screen.line()
+                        screen.print()
+                        with open(f".lifelog/{username}/{titles[choice].replace(' ','_').lower()}.dat", 'wb') as file:
+                            data = { "content": diary, "mood": newMood, "date": datetime.now().strftime("%d %B %Y") }
+                            pickle.dump(data, file)
+                        screen.line()
+                        screen.print("[✔] Entry updated successfully!", color=COLORS.LIGHT_GREEN)
+                        screen.input("Press ENTER to go back", color=COLORS.DARK_GRAY)
+                        screen.clear()
+                        screen.render()
+
+                elif action == 2:
+                    confirm = screen.input("Are you sure you want to delete this entry? (Y/N)", color=COLORS.LIGHT_RED)
+                    if confirm.upper() == 'Y':
+                        os.remove(f'.lifelog/{username}/{titles[choice].replace(" ", "_").lower()}.dat')
+                    screen.clear()
+                    screen.render()
+                else:
+                    screen.clear()
+                    screen.render()
+        except Exception:
+            screen.clear()
+            screen.render()
     
 
 def Dashboard(username: str):
@@ -422,34 +457,51 @@ def AddEntry(username: str):
     Header("✍️  ADD A NEW ENTRY  ✍️")
     screen.print(f"Date: {datetime.now().strftime('%d %B %Y')} | User: {username}", color=COLORS.YELLOW)
     title = screen.input("Title: ", color=COLORS.YELLOW)
-    screen.line()
-    screen.print("Write your diary below (Type END on a new line to finish)", color=COLORS.DARK_GRAY)
-    screen.line()
-    diary = screen.editor("> ", color=COLORS.DARK_GRAY) 
-    screen.print()
-    screen.line()
-    screen.print("[1] 😊 Happy")
-    screen.print("[2] 😌 Calm")
-    screen.print("[3] 😔 Sad")
-    screen.print("[4] 😤 Stressed")
-    screen.print("[5] 😎 Confident")
-    screen.line()
-    mood = screen.input("Enter your mood (1-5):", color=COLORS.YELLOW)
-    screen.line()
-    if mood == '':
-        mood = '1'
+    if title.strip().lower().replace(" ", "_") + ".dat" in os.listdir(f".lifelog/{username}/"):
+        screen.print("[X] You already have a diary with the same name. Choose a different name", color=COLORS.LIGHT_RED)
+        screen.render()
+        time.sleep(1)
+        screen.clear()
+        AddEntry(username)
+    elif len(title.strip()) == 0:
+        screen.print("[X] Title cannot be empty.", color=COLORS.LIGHT_RED)
+        screen.render()
+        time.sleep(1)
+        screen.clear()
+        AddEntry(username)
+    else:
+        screen.line()
+        screen.print("Write your diary below (Type END on a new line to finish)", color=COLORS.DARK_GRAY)
+        screen.line()
+        diary = screen.editor("> ", color=COLORS.DARK_GRAY) 
+        screen.print()
+        screen.line()
+        screen.print("[1] 😊 Happy")
+        screen.print("[2] 😌 Calm")
+        screen.print("[3] 😔 Sad")
+        screen.print("[4] 😤 Stressed")
+        screen.print("[5] 😎 Confident")
+        screen.line()
+        mood = screen.input("Enter your mood (1-5):", color=COLORS.YELLOW)
+        screen.line()
+        if not mood.isnumeric() or int(mood) < 0 or int(mood) > 5:
+            screen.print("[X] Invalid mood", color=COLORS.LIGHT_RED)
+            screen.print("Selecting [1] Happy by default. (You can change this later)", color=COLORS.DARK_GRAY)
+            mood = '1'
+        if mood == '':
+            mood = '1'
 
-    with open(f".lifelog/{username}/{title.replace(' ','_').lower()}.dat", 'wb') as file:
-        data = { "content": diary, "mood": mood, "date": datetime.now().strftime("%d %B %Y") }
-        pickle.dump(data, file)
+        with open(f".lifelog/{username}/{title.replace(' ','_').lower()}.dat", 'wb') as file:
+            data = { "content": diary, "mood": mood, "date": datetime.now().strftime("%d %B %Y") }
+            pickle.dump(data, file)
 
-    screen.print()
-    screen.print("""Saving your entry...
-[✔] Entry saved successfully! 🌿
-“Remember, your thoughts are seeds — keep planting good ones.”""", color=COLORS.LIGHT_GREEN)
-    screen.line()
-    screen.input("Press ENTER to continue: ", color=COLORS.DARK_GRAY)
-    screen.clear()
+        screen.print()
+        screen.print("""Saving your entry...
+    [✔] Entry saved successfully! 🌿
+    “Remember, your thoughts are seeds — keep planting good ones.”""", color=COLORS.LIGHT_GREEN)
+        screen.line()
+        screen.input("Press ENTER to continue: ", color=COLORS.DARK_GRAY)
+        screen.clear()
 
 def CreateUser():
     users = []
@@ -644,8 +696,8 @@ $$/      $$/  $$$$$$$/ $$/  $$$$$$$/  $$$$$$/  $$/  $$/  $$/  $$$$$$$/
                     stats = loadStats()
                     Dashboard(stats['username'])
         elif choice == '2':
-            Header("LifeLog Docs")
-            screen.input("Press ENTER to go back!")
+            ShowHelp(screen)
+            screen.input("Press ENTER to go back", color=COLORS.DARK_GRAY)
 
         elif choice == '3':
             screen.clear()
